Changed Variable Scope for Command-Line Args
Changed the variable scope from 'my' to 'our' in process_options as it was
otherwise not honoring the command-line arguments and, instead, using the
built-in.

Also changed the defaults to something more conservative. Warning threshold
is now 80% and critical is 95%.

Default Perl and Nagios paths were also updated to reflect a system install.